requirements.txt
langchain
langchain-huggingface
llama-cpp-python
sentence-transformers
pinecone-client
flask
pypdf
python-dotenv
-e .